fix (Footer) rename Footer.tsx to FooterComponent.tsx
